move communicator helpers into communicator dir
