+++ builder/proxmox/clone/step_ssh_key_pair.go
@@ -6,7 +6,7 @@ import (
 	"os"
 
 	common "github.com/hashicorp/packer/builder/proxmox/common"
-	"github.com/hashicorp/packer/helper/ssh"
+	"github.com/hashicorp/packer/helper/communicator/ssh"
 	"github.com/hashicorp/packer/packer"
 	"github.com/hashicorp/packer/packer-plugin-sdk/multistep"
 	"github.com/hashicorp/packer/packer-plugin-sdk/uuid"
