+++ builder/virtualbox/common/step_ssh_key_pair.go
@@ -6,7 +6,7 @@ import (
 	"os"
 
 	"github.com/hashicorp/packer/helper/communicator"
-	"github.com/hashicorp/packer/helper/ssh"
+	"github.com/hashicorp/packer/helper/communicator/ssh"
 	"github.com/hashicorp/packer/packer"
 	"github.com/hashicorp/packer/packer-plugin-sdk/multistep"
 	"github.com/hashicorp/packer/packer-plugin-sdk/uuid"
