+++ builder/vmware/common/driver_esx5.go
@@ -29,7 +29,7 @@ import (
 
 	"github.com/hashicorp/go-getter/v2"
 	"github.com/hashicorp/packer/helper/communicator"
-	helperssh "github.com/hashicorp/packer/helper/ssh"
+	helperssh "github.com/hashicorp/packer/helper/communicator/ssh"
 	"github.com/hashicorp/packer/packer"
 	"github.com/hashicorp/packer/packer-plugin-sdk/multistep"
 	"github.com/hashicorp/packer/packer-plugin-sdk/sdk-internals/communicator/ssh"
